requirements.txt
mouse
pillow